requirements.txt
# Core Dependencies
boto3>=1.34.0
google-auth>=2.27.0
gspread>=5.12.0
python-dotenv>=1.0.0
python-telegram-bot>=20.7

# ML Dependencies
lightgbm>=4.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
joblib>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
